moved last ditch effort from issue 8 to the end of the route find method, as that was short circuiting the issue discovered in issue22.
