+++ tree_test.go
@@ -3,29 +3,3 @@
 // can be found in the LICENSE file included.
 
 package vestigo
-
-import "fmt"
-
-// printTree - Helper method to print a representation of the tree
-func (n *node) printTree(pfx string, tail bool) {
-	p := prefix(tail, pfx, "└── ", "├── ")
-	fmt.Printf("%s%s, %p: type=%d, parent=%p, resource=%v\n", p, n.prefix, n, n.typ, n.parent, n.resource)
-
-	children := n.children
-	l := len(children)
-	p = prefix(tail, pfx, "    ", "│   ")
-	for i := 0; i < l-1; i++ {
-		children[i].printTree(p, false)
-	}
-	if l > 0 {
-		children[l-1].printTree(p, true)
-	}
-}
-
-// prefix - print the prefix
-func prefix(tail bool, p, on, off string) string {
-	if tail {
-		return fmt.Sprintf("%s%s", p, on)
-	}
-	return fmt.Sprintf("%s%s", p, off)
-}
